Split oneway.rs (518 lines) into oneway/ subdirectory
Split crates/simulation/src/oneway.rs into focused modules under oneway/:
- types.rs: OneWayDirection, OneWayDirectionMap, ToggleOneWayEvent
- systems.rs: handle_toggle_oneway, rebuild_csr_with_oneway, CsrGraph::from_road_network_filtered
- persistence.rs: Saveable impl for OneWayDirectionMap
- plugin.rs: OneWayPlugin
- tests.rs: all unit tests
- mod.rs: thin facade with re-exports

All public items re-exported from mod.rs — no external import changes needed.
